+++ spec/lib/nquery/configuration_spec.rb
@@ -86,6 +86,6 @@ end
 
 RSpec.describe Nquery::VERSION do
   it "defines the gem version" do
-    expect(Nquery::VERSION).to eq("0.1.2")
+    expect(Nquery::VERSION).to eq("0.1.3")
   end
 end